Move UAST-based Lint rules to dedicated package
