+++ lint/src/test/java/permissions/dispatcher/PermissionsDispatcherIssueRegistryTest.java
@@ -1,6 +1,9 @@
 package permissions.dispatcher;
 
 import org.junit.Test;
+import permissions.dispatcher.uast.UastCallNeedsPermissionDetector;
+import permissions.dispatcher.uast.UastCallOnRequestPermissionsResultDetector;
+import permissions.dispatcher.uast.UastNoCorrespondingNeedsPermissionDetector;
 
 import static com.google.common.truth.Truth.assertThat;
 
@@ -8,8 +11,8 @@ public final class PermissionsDispatcherIssueRegistryTest {
     @Test
     public void issues() {
         assertThat(new PermissionsDispatcherIssueRegistry().getIssues()).containsAllOf(
-                CallNeedsPermissionDetector.ISSUE,
-                CallOnRequestPermissionsResultDetector.ISSUE,
-                NoCorrespondingNeedsPermissionDetector.ISSUE);
+                UastCallNeedsPermissionDetector.ISSUE,
+                UastCallOnRequestPermissionsResultDetector.ISSUE,
+                UastNoCorrespondingNeedsPermissionDetector.ISSUE);
     }
 }
