+++ lint/src/test/java/permissions/dispatcher/UastNoCorrespondingNeedsPermissionDetectorTest.java
@@ -2,6 +2,7 @@ package permissions.dispatcher;
 
 import org.intellij.lang.annotations.Language;
 import org.junit.Test;
+import permissions.dispatcher.uast.UastNoCorrespondingNeedsPermissionDetector;
 
 import static com.android.tools.lint.checks.infrastructure.TestFiles.java;
 import static com.android.tools.lint.checks.infrastructure.TestLintTask.lint;
@@ -10,7 +11,7 @@ import static permissions.dispatcher.Utils.SOURCE_PATH;
 import static permissions.dispatcher.Utils.getOnNeedsPermission;
 import static permissions.dispatcher.Utils.getOnRationaleAnnotation;
 
-public final class NoCorrespondingNeedsPermissionDetectorTest {
+public final class UastNoCorrespondingNeedsPermissionDetectorTest {
 
     @Test
     public void noNeedsPermissionAnnotationNoErrors() throws Exception {
@@ -34,7 +35,7 @@ public final class NoCorrespondingNeedsPermissionDetectorTest {
                         java(SOURCE_PATH + "NeedsPermission.java", onNeeds),
                         java(SOURCE_PATH + "OnShowRationale.java", onShow),
                         java(SOURCE_PATH + "Foo.java", foo))
-                .issues(NoCorrespondingNeedsPermissionDetector.ISSUE)
+                .issues(UastNoCorrespondingNeedsPermissionDetector.ISSUE)
                 .run()
                 .expectClean();
     }
@@ -55,7 +56,7 @@ public final class NoCorrespondingNeedsPermissionDetectorTest {
         String expectedText = ""
                 + SOURCE_PATH + "Foo.java:3: Error: Useless @OnShowRationale declaration "
                 + "["
-                + NoCorrespondingNeedsPermissionDetector.ISSUE.getId()
+                + UastNoCorrespondingNeedsPermissionDetector.ISSUE.getId()
                 + "]\n"
                 + "@OnShowRationale(\"Camera\")\n"
                 + "~~~~~~~~~~~~~~~~~~~~~~~~~~\n"
@@ -65,7 +66,7 @@ public final class NoCorrespondingNeedsPermissionDetectorTest {
                 .files(
                         java(SOURCE_PATH + "OnShowRationale.java", onShow),
                         java(SOURCE_PATH + "Foo.java", foo))
-                .issues(NoCorrespondingNeedsPermissionDetector.ISSUE)
+                .issues(UastNoCorrespondingNeedsPermissionDetector.ISSUE)
                 .run()
                 .expect(expectedText)
                 .expectErrorCount(1)
